6-Commit [viernes_19-06-2015] Nueva distribución en el ménu y en la forma de acceder al sistema.

+++ GrupoEmpresarial-war/src/java/com/dane/ige/utilidad/Ventana.java
@@ -8,14 +8,16 @@ import org.primefaces.context.RequestContext;
 
 /**
  * Clase que administra los métodos y propiedades de una ventana modal.
+ *
  * @author srojasm
- */ 
+ */
 @ManagedBean(name = "dfView")
-public class Ventana implements Serializable{
- 
+public class Ventana implements Serializable {
+
     /**
      * Método que permite visualizar un componente en una ventana modal.
-     * @param componente 
+     *
+     * @param componente
      */
     public void visualizarVentana(String componente) {
         RequestContext.getCurrentInstance().openDialog(componente);
@@ -24,21 +26,22 @@ public class Ventana implements Serializable{
     /**
      * Método que permite visualizar un componente en una ventana modal, pasando
      * los parametros de modal, dragable, resizable y la altitura.
+     *
      * @param componente
      * @param modal
      * @param dragable
      * @param resizable
-     * @param alto 
-     * @param ancho 
+     * @param alto
+     * @param ancho
      */
-    public void visualizarVentanaParametrizada(String componente, boolean modal, boolean dragable, boolean resizable, int alto, int ancho) {
-        Map<String,Object> options = new HashMap<String, Object>();
+    public void visualizarVentanaParametrizada(String componente, boolean modal, boolean dragable, boolean resizable, Long alto, Long ancho) {
+        Map<String, Object> options = new HashMap<String, Object>();
         options.put("modal", modal);
         options.put("draggable", dragable);
         options.put("resizable", resizable);
         options.put("contentHeight", alto);
         options.put("contentWidth", ancho);
-        RequestContext.getCurrentInstance().openDialog(componente, options, null);        
+        RequestContext.getCurrentInstance().openDialog(componente, options, null);
     }
-    
+
 }
